added drawing styles preview; bugfixes
