+++ modules/AIDEAdmin/app.py
@@ -8,6 +8,7 @@
 
 import os
 import html
+from threading import Thread
 from bottle import SimpleTemplate, request, redirect, abort
 from constants.version import AIDE_VERSION
 from .backend.middleware import AdminMiddleware
@@ -25,10 +26,7 @@ class AIDEAdmin:
         self.middleware = AdminMiddleware(config)
 
         # ping connected AIController, FileServer, etc. servers and check version
-        try:
-            self.middleware.getServiceDetails(True)
-        except Exception as e:
-            pass
+        Thread(target=self._queryServiceDetails).start()
 
         self._initBottle()
     
@@ -40,6 +38,14 @@ class AIDEAdmin:
     def addLoginCheckFun(self, loginCheckFun):
         self.login_check = loginCheckFun
 
+    
+    def _queryServiceDetails(self):
+        try:
+            self.middleware.getServiceDetails(True)
+        except:
+            # not running on main host; ignore error
+            pass
+    
 
     def _initBottle(self):
