+++ modules/LabelUI/static/templates/interface.html
@@ -9,6 +9,7 @@
     <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js?v={{ version }}" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+    <script type="text/javascript" src="/static/interface/js/util.js?v={{ version }}"></script>
     <script type="text/javascript" src="/static/interface/js/uiControlHandler.js?v={{ version }}"></script>
     <script type="text/javascript" src="/static/interface/js/commandListener.js?v={{ version }}"></script>
     <script type="text/javascript" src="/static/interface/js/configuration.js?v={{ version }}"></script>
