+++ modules/LabelUI/static/templates/tutorial_welcome.html
@@ -106,31 +106,7 @@
             error: function(xhr, status, error) {
                 //TODO
             }
-        })
-
-        // // create example predictions/annotations
-        // var predVP = new ImageViewport($('#uitutorial-predictions'), true);
-        // var predStyle = JSON.parse(JSON.stringify(window.styles.predictions));
-        // var color = '#7AD124';
-        // var className = 'Hartebeest';
-        // predStyle['strokeColor'] = window.addAlpha(color, predStyle.lineOpacity);
-        // predStyle['fillColor'] = window.addAlpha(color, predStyle.fillOpacity);
-        // predVP.addRenderElement(new ImageElement('img', predVP, '/static/interface/exampleData/sample_image.jpg', 0, true));
-
-        // var annoVP = new ImageViewport($('#uitutorial-annotations'), true);
-        // var annoStyle = JSON.parse(JSON.stringify(window.styles.annotations));
-        // annoStyle['strokeColor'] = window.addAlpha(color, annoStyle.lineOpacity);
-        // annoStyle['fillColor'] = window.addAlpha(color, annoStyle.fillOpacity);
-        // annoVP.addRenderElement(new ImageElement('img', annoVP, '/static/interface/exampleData/sample_image.jpg', 0, true));
-
-
-        // if(window.predictionType === 'labels') {
-        //     predVP.addRenderElement(new BorderStrokeElement('predLabel', className, predStyle, false, 1, true));
-        // } else if(window.predictionType === 'points') {
-        //     predVP.addRenderElement(new PointElement('predPoint', 0.5, 0.55, predStyle, false, 1, true));
-        // } else if(window.predictionType === 'boundingBoxes') {
-        //     predVP.addRenderElement(new RectangleElement('predBox', 0.5, 0.55, 0.5, 0.55, predStyle, false, 1, true));
-        // }
+        });
 
         var taskDescr = 'providing labels';
         if(window.annotationType === 'labels') {
