+++ modules/ProjectAdministration/static/templates/projectConfiguration.html
@@ -10,6 +10,14 @@
         <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
         <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+        <!-- TODO: ugly solution; these are only needed in "userInterface.html", but should not be loaded twice: -->
+        <script type="text/javascript" src="/static/interface/js/labelClassHandler.js?v={{ version }}"></script>
+        <script type="text/javascript" src="/static/interface/js/dataEntry.js?v={{ version }}"></script>
+        <script type="text/javascript" src="/static/interface/js/annotationPrimitives.js?v={{ version }}"></script>
+        <script type="text/javascript" src="/static/interface/js/viewport/imageViewport.js?v={{ version }}"></script>
+        <script type="text/javascript" src="/static/interface/js/renderPrimitives.js?v={{ version }}"></script>
+        <script type="text/javascript" src="/static/interface/js/util.js?v={{ version }}"></script>
+        <!-- /TODO -->
         <script src="/static/general/js/overlay.js?v={{ version }}"></script>
         <script type="text/javascript">
             function parseURLparameter(name) {
